requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
PySocks==1.7.1
cryptography==42.0.5